xong admin

+++ src/main/java/com/websiteshop/dao/ProductDAO.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import com.websiteshop.entity.Product;
 
 @Service
-public interface ProductDAO extends JpaRepository<Product, Integer> {
+public interface ProductDAO extends JpaRepository<Product, Long> {
       @Query("SELECT p FROM Product p WHERE p.category.categoryId=?1")
       List<Product> findByCategoryId(String cid);
 }
